+++ app/FileManager/FileManager.php
@@ -76,13 +76,19 @@ class FileManager
         return false !== file_put_contents($file, $contents);
     }
 
-    public function delete($file): bool
+    public function delete($path)
     {
-        if (!file_exists($file)) {
-            return true;
+        $remove = is_dir($path) ? 'rmdir' : 'unlink';
+        $error = ['code' => 500, 'msg' => 'Failed removing' . $path];
+
+        if (!file_exists($path)) {
+            return ['error' => null];
+        }
+        if (!is_writable($path)) {
+            return ['error' => ['code' => 403, 'msg' => 'No permission to write path']];
         }
 
-        return is_dir($file) ? false : unlink($file);
+        return ['error' => $remove($path) ? null : $error];
     }
 
     private function loadFilePermissions(string $path): array